creating separate actions to set viewerDocumentToDisplay for messages and case detail view
