+++ web-client/src/presenter/actions/setMessageDetailViewerDocumentToDisplayAction.js
@@ -1,0 +1,45 @@
+import { formatAttachments } from '../../../../shared/src/business/utilities/formatAttachments';
+import { state } from 'cerebral';
+
+/**
+ * sets the message detail viewerDocumentToDisplay from props
+ *
+ * @param {object} providers the providers object
+ * @param {object} providers.get the cerebral get method
+ * @param {object} providers.props the cerebral props object
+ * @param {object} providers.store the cerebral store object
+ */
+export const setMessageDetailViewerDocumentToDisplayAction = async ({
+  applicationContext,
+  get,
+  props,
+  store,
+}) => {
+  const { mostRecentMessage, viewerDocumentToDisplay } = props;
+  const caseDetail = get(state.caseDetail);
+  const { docketNumber } = caseDetail;
+
+  store.set(state.viewerDocumentToDisplay, viewerDocumentToDisplay);
+
+  const { attachments } = mostRecentMessage;
+  const formattedAttachments = formatAttachments({ attachments, caseDetail });
+  console.log(formattedAttachments);
+  const formattedAttachment = formattedAttachments.find(
+    attachment => attachment.documentId === viewerDocumentToDisplay.documentId,
+  );
+
+  if (viewerDocumentToDisplay && !formattedAttachment.archived) {
+    const {
+      url,
+    } = await applicationContext
+      .getUseCases()
+      .getDocumentDownloadUrlInteractor({
+        applicationContext,
+        docketNumber,
+        documentId: viewerDocumentToDisplay.documentId,
+        isPublic: false,
+      });
+
+    store.set(state.iframeSrc, url);
+  }
+};
