--- web-client/src/presenter/actions/setViewerDocumentToDisplayAction.test.js
@@ -1,82 +1,0 @@
-import { applicationContextForClient as applicationContext } from '../../../../shared/src/business/test/createTestApplicationContext';
-import { presenter } from '../presenter-mock';
-import { runAction } from 'cerebral/test';
-import { setViewerDocumentToDisplayAction } from './setViewerDocumentToDisplayAction';
-
-describe('setViewerDocumentToDisplayAction', () => {
-  beforeAll(() => {
-    applicationContext
-      .getUseCases()
-      .getDocumentDownloadUrlInteractor.mockReturnValue({
-        url: 'www.example.com',
-      });
-    presenter.providers.applicationContext = applicationContext;
-  });
-
-  it('sets the viewerDocumentToDisplay from props on state and sets the iframeSrc url from the return from the use case', async () => {
-    const result = await runAction(setViewerDocumentToDisplayAction, {
-      modules: {
-        presenter,
-      },
-      props: {
-        mostRecentMessage: { attachments: [] },
-        viewerDocumentToDisplay: { documentId: '1234' },
-      },
-      state: {
-        caseDetail: {
-          docketNumber: '123-45',
-          documents: [],
-        },
-        viewerDocumentToDisplay: null,
-      },
-    });
-
-    expect(result.state.viewerDocumentToDisplay).toEqual({
-      documentId: '1234',
-    });
-    expect(result.state.iframeSrc).toEqual('www.example.com');
-  });
-
-  it('does not set iframeSrc if props.viewerDocumentToDisplay is null', async () => {
-    const result = await runAction(setViewerDocumentToDisplayAction, {
-      modules: {
-        presenter,
-      },
-      props: {
-        mostRecentMessage: { attachments: [] },
-        viewerDocumentToDisplay: null,
-      },
-      state: {
-        caseDetail: {
-          docketNumber: '123-45',
-          documents: [],
-        },
-        viewerDocumentToDisplay: null,
-      },
-    });
-
-    expect(result.state.iframeSrc).toBeUndefined();
-  });
-
-  it('does not set iframeSrc when the document to display has been archived', async () => {
-    const result = await runAction(setViewerDocumentToDisplayAction, {
-      modules: {
-        presenter,
-      },
-      props: {
-        mostRecentMessage: { attachments: [{ documentId: '1234' }] },
-        viewerDocumentToDisplay: { documentId: '1234' },
-      },
-      state: {
-        caseDetail: {
-          archivedDocuments: [{ archived: true, documentId: '1234' }],
-          docketNumber: '123-45',
-          documents: [],
-        },
-        viewerDocumentToDisplay: null,
-      },
-    });
-
-    expect(result.state.iframeSrc).toBeUndefined();
-  });
-});
