+++ web-client/src/presenter/sequences/changeTabAndSetViewerDocumentToDisplaySequence.js
@@ -1,4 +1,3 @@
-import { getMostRecentMessageInThreadAction } from '../actions/getMostRecentMessageInThreadAction';
 import { setDefaultCaseDetailTabAction } from '../actions/setDefaultCaseDetailTabAction';
 import { setIsPrimaryTabAction } from '../actions/setIsPrimaryTabAction';
 import { setViewerDocumentToDisplayAction } from '../actions/setViewerDocumentToDisplayAction';
@@ -6,6 +5,5 @@ import { setViewerDocumentToDisplayAction } from '../actions/setViewerDocumentTo
 export const changeTabAndSetViewerDocumentToDisplaySequence = [
   setDefaultCaseDetailTabAction,
   setIsPrimaryTabAction,
-  getMostRecentMessageInThreadAction,
   setViewerDocumentToDisplayAction,
 ];
