+++ web-client/src/presenter/sequences/setMessageDetailViewerDocumentToDisplaySequence.js
@@ -1,0 +1,10 @@
+import { getMostRecentMessageInThreadAction } from '../actions/getMostRecentMessageInThreadAction';
+import { setMessageDetailViewerDocumentToDisplayAction } from '../actions/setMessageDetailViewerDocumentToDisplayAction';
+import { showProgressSequenceDecorator } from '../utilities/sequenceHelpers';
+
+export const setMessageDetailViewerDocumentToDisplaySequence = showProgressSequenceDecorator(
+  [
+    getMostRecentMessageInThreadAction,
+    setMessageDetailViewerDocumentToDisplayAction,
+  ],
+);
