+++ web-client/src/presenter/sequences/setViewerDocumentToDisplaySequence.js
@@ -1,7 +1,6 @@
-import { getMostRecentMessageInThreadAction } from '../actions/getMostRecentMessageInThreadAction';
 import { setViewerDocumentToDisplayAction } from '../actions/setViewerDocumentToDisplayAction';
 import { showProgressSequenceDecorator } from '../utilities/sequenceHelpers';
 
 export const setViewerDocumentToDisplaySequence = showProgressSequenceDecorator(
-  [getMostRecentMessageInThreadAction, setViewerDocumentToDisplayAction],
+  [setViewerDocumentToDisplayAction],
 );
